refactor: todoTitle, TodoList component folder refactoring TodoStore refactoring
