+++ holee-todo/src/App.css
@@ -1,38 +1,18 @@
-.App {
-  text-align: center;
+.todo-item-content {
+  background-color: white;
+  border: 0;
+  outline: 0;
+  font-weight: bold;
+  font-size: 16px;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+.todo-item-checked {
+  text-decoration: line-through;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.todo-item-remove-button {
+  background-color: white;
+  padding-left: 8px;
+  border: 0;
+  outline: 0;
 }
