+++ holee-todo/src/component/Title/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TodoTitle';
