+++ holee-todo/src/component/TodoItem/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TodoItem';
