+++ holee-todo/src/component/TodoList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TodoList';
