+++ holee-todo/src/stores/TodoStore.tsx
@@ -22,10 +22,14 @@ class TodoStore {
     return targetId !== -1;
   };
 
+  removeTodoItem = (targetId: number) => {
+    this.todoList.splice(targetId, 1);
+  };
+
   removeContent = (id: number) => {
     const targetId = this.todoList.findIndex((v) => v.id === id);
     if (this.checkId(targetId)) {
-      this.todoList.splice(targetId, 1);
+      this.removeTodoItem(targetId);
     }
   };
 }
